Correctly map representing and self representing properties
The OPTICS system uses a dropdown menu in order to set the correct human
readable value for each case. This means that it expects a single
alphabetic code to be the value in the payload for the service type,
query type and applicant type properties.

Also what is sent in the payload differs greatly dependent upon whether
the person filling out the form is a representative or not.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>