Add ClaimNumber property to OPTICS payload replacing NewOrExistingClaim
HMCTS have updated their form pages to add an additional property which
is sent in the Submission payload: `ClaimNumber`.

This has two potential values: `CaseReferenceYes` and `CaseReferenceNo`.

Funcationally wise this replaces the previous logic around
`NewOrExistingClaim` property. This property will still exist in the
payload, as their form will pass it on to the adapter, however it is not
used at all. I have left that in the test payloads because it is still
received, just never mapped.