Add Assigned Team to Complaints payload
The HMCTS team would like to have the complaint location value mapped
to another field. In production that field is AssignedTeam and then in
staging that field is AssignedTeamSS.

The adapter has no knowledge of the environment it is running in
currently so the quickest way is to add both fields to the payload and
OPTICS will pick the field it needs upon receipt.