Add submission ID to all OPTICS payloads
In order for the adapter to be able to check whether OPTICS has received
a submission previously we need to send the submission ID for every
submission in the payload. We do this by adding it as an ExternalId.
Optics will then save this as an additional property in their database.