Update all jobs to check for existence of previous submissions
Before allowing each job to post the submission payload to the OPTICS we
check whether OPTICS has previously received that a payload containing
the submission id. We do this by requesting an attribute for a specific
case passing the submission ID as the ExternalId URL parameter.

OPTICS will return a 200 if it has a case which has that submission ID
as an ExternalId. If this is the case we will then add a
ProcessedSubmission record with that submission ID to the DB.

If it does not have a case with that ExternalId then it will return a
400. In this event, as a case with that submission ID does not exist in
OPTICS, we want to post the submission in order to create the case as
normal.

In the event of a networking error communicating with OPTICS to check
for a previously processed submission we fail and the job should retry
a little later.