Add prometheus rules (moved from fb-monitoring)
These rules were in the `fb-monitoring` and applied there with custom scripts and the need to have 2 eks tokens one for staging and another for production.

It seems more appropriate to have these rules in the application repo.

If this works, the rules will be deleted from the fb-monitoring repo.