Support both types of address
We now have a concrete address component, where before forms had to implement their own address page with multiple input texts.

This address component is a bit different as it propagates the different parts as JSON.

This change makes it support both type of address for a while until all forms are migrated.