Cleanup legacy address code
https://trello.com/c/RfKx3plj

After the introduction of the new address component, we implemented some backwards compatible code to support both types of addresses for a while until the complaints form was fully migrated.

Now the form (English and Welsh versions) is migrated, we can delete the backwards compatibility code and streamline the tests.